Reuse details pane
